Make stats-log-api-gen available as a prebuilt
This involved two steps. First, we make stats-log-api-gen statically
link all of its dependencies. This prevents needing to pull in
prebuilts from frameworks/base, for example.

Second, we add a module_exports entry to Android.bp, which will
actually make stats-log-api-gen available as a prebuilt to modules
that need it.

Bug: 178500083
Test: m stats-log-api-gen-exports
Change-Id: I2dfa215368beeeab73de7732596ff38715d64542