Update patch set 1
Patch Set 1:

Change bypassed Presubmit.

Bypass Reason: other 
Additional Feedback: TH was complaining about a merge conflict, however this CL was a cherry-pick of internal change ag/13044249 so it shouldn't merge back to internal. Due to shuffling of file locations in internal tree, it was not possible to use Merged-In clause to skip the automerging, so I will manually skip the merge conflict once automerger attempts to merge this to sc-dev.

Patch-set: 1
Tag: autogenerated:TreeHugger:Bypassed-Presubmit