Add new atom DataUsageBytesTransferV2
This is a new atom for unmetered 5g metrics. In current design,
the value metric does not support aggregating across dimensions
on a pull so there will have problem if just adding a new field
to DataUsageBytesTransfer. Therefore, creating a new atom and it
is a copy from DataUsageBytesTransfer and add a is_metered field.

Bug: 183776809
Test: build pass
Change-Id: Ieac300029cbb42b447314905cc22d2ede7f1669c
Merged-In: Ieac300029cbb42b447314905cc22d2ede7f1669c