Remove METERED_ALL and DSS_ALL states from enum of DataUsageBytesTransferV2
Original idea is to make this atom more flexible, if someone
needs new dimension but doesn't care about the meteredness or
opportunistic. It can extend this atom, instead of creating a
new one with one extra field. Thus the atom definition could
be reused, just fill the combined state METERED_ALL or DSS_ALL
to the corresponding field, and aggregate stats over meteredness
or opportunistic at device side to reduce the upload amount.
However, sharing the atom definition proto across multiple atoms
is not supported. The idea is infeasible in current design, so
removing METERED_ALL and DSS_ALL from the enum.

This change also addresses comments from ag/202486465.
Removing the comments that is not true for the atoms.

Bug: 183776809
Bug: 202486465
Test: build pass
Change-Id: I8b9fbac9f096f187f445e62324c9f1ca008f0faf