Revert "Update the documentation of some `OdrefreshReported` fields."
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/6bda4e07-cc0f-4d2a-8470-b6ae445286fd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I87c78e40b36793ecae5fc284c8be4c12bfb4fe78