Revert "Adding the metric's atom of NetworkDnsServerSupportReported"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/2d60fa9f-a529-447a-8ca7-15d4522c65de). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ided1ed15cc5443437bb07dffea9d30b4ebeab15d