Revert "Add UT_LOCAL_ONLY in enum UpstreamType."
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/d6592e60-fe57-4125-853f-04711e2edb0a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I492a8724060a6726adce43665ba5258a0d74019d