Revert "[TeX] Added Counter atom with UID dimension"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ab7f73c0-33e9-48d0-81b8-5aca2c284996). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I88a957f11d7762c37c0ee7a67fc7ac70061a639f