Add Cronet's atoms to AOSP
Cherry-pick to branches required for external/cronet development

Original commit message:
Now that Cronet is shipped as part of the Tethering mainline module,
it makes more sense to move its atoms IDs in the platform range.

The atoms here have been copied 1:1 from
https://source.corp.google.com/search?q=f:cronet_atoms.proto.
If copied downstream as is, these would clash with the pre-existing
ones in the app range (since the extension and proto field will have
the same name).
To solve this issue, we will deprecate the existing atoms in the app range
and have Cronet always log using the IDs of this extension.

Test: I've checked statsd output for a local build with logging enabled
Bug: 284972285

Merged-In: Ie69d2c14c3635bb4c41f25d6da6ffe622ba6dd00
Change-Id: I7c6d3db01e70b36bab6b1a33c9d75e902391022a