Revert "Adding Counter metrics for missing UUIDs, fetching UUIDs..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4248558b-a9e3-4183-afcc-7cd09a27a582). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia1bffe36f9bc526898ea210d474b00758cc16384