Revert "Add enum items for screen on/off event counter metrics"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b47e2b84-4f54-48f2-810e-658e1debee2d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0b7c6db7a9f4bbcfb02f19dd5a80b39233d68ba5