Revert "Move ART-related atoms to the extensions model"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a826eef2-e4ff-48e6-97be-114ff2f28a76). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I747edeff9aef7e3549b5756fb1c4f493236df282