Revert "media codec usage pushed atoms"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e326738a-5ebc-4d3f-bd0b-4b5e9771f28f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I31307e8d44f74dae279bceacc04e62a47e28c090