Revert "media codec usage pushed atoms"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/beeb60aa-4f17-4248-b6bd-cfb536895a55). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ieb41fde33c1c4b33d3fbd2254fe16c7b50ce4c37