Revert "app concurrent codec usage"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/beeb60aa-4f17-4248-b6bd-cfb536895a55). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If8c4c3a5b158f6bca80dce2b7c811d9678d5db90