Revert "Add CMC background GC Datum"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/35ff8110-0017-49b4-83d4-37033438e758). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I63b223d5d1704e193d70060f643accc643d95b68