Revert "Add CMC background GC Datum"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/92f2b0ad-d96a-48ad-8a2d-898d243e131a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iffc9c46dcff7b976f49f2806e746488ef006c221