Revert "Add a TEX metric to track an invalid call to BluetoothAd..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ce97cd35-096f-437a-84b5-855e3ba440ea). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I02e76b48d4153fd2d7ea072176d4bab75966c03d