Revert "Add a TEX metric to track an invalid call to BluetoothAd..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e8a1db77-1950-4100-9655-24d85668c996). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7d81eaca4b27e84602b943e14e7b623dd0897b4e