Revert "Add a TEX metric to track an invalid call to BluetoothAd..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6292ad6a-0558-4a02-a684-8755603c0bae). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie99c7905dfa60d688683d24b66aae5b789f238f1