Revert "Add string field to RescuePartyResetReported atom"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b2760a20-b944-42fc-92f6-d1b08fc7d9e5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iec784c348906a45835d4f99429836890cb946328