Revert "Add @RequiresApi annotation to generated code for repeat..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ea78e2b7-db38-4c6c-9062-2705dde63e4b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I18db42f682d53d919e4cc4b0bb2c75e342375db5