Revert "Add @RequiresApi annotation to generated code for repeat..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d7a40d3d-e372-4ac7-8e7f-bb14c727eaf5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibfff21d1a677b56e827d5c3c53f86db16166d7f5