Revert "Add @RequiresApi annotation to generated code for repeat..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/70daf2cf-9764-476e-ae72-b52c5384504b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9cd1bb897f002b63775bdc5f7b36676ad218d7d3