Add a quirk event for the DHCPv6 heuristic being triggered
Currently DHCPv6 is only started when the network provides a default route,
but no address via autoconfiguration.
This is arguably a network quirk (albeit not necessarily in a negative
sense). Having this metric will give us a sense on how often the DHCPv6-PD
heuristic is triggered and therefore might let us gauge the potential impact of enabling the DHCPv6-PD experiment for the entire ecosystem.

Change-Id: Ia739ba6e9c58a4c0efbf582f67c62dd78ff44319
Test: none