Revert "Defining __ANDROID_API_T__ and __INTRODUCED_IN on non-An..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e9b8ae6d-7e17-4cd5-8a97-edfc3d06dd7f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0603fab9edb0fcf10226b799bef85ca69d5dbe52
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com