Revert "Adding __INTRODUCED_IN to stats_write if contains repeat..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e9b8ae6d-7e17-4cd5-8a97-edfc3d06dd7f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib4f800aa04c11660540087a9dcf59ca6a3306691
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com