Revert "Defining __ANDROID_API_T__ and __INTRODUCED_IN on non-An..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c4883a9e-fb23-4457-92a8-a5eef84088dc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 353432148
Change-Id: I447539a1e138b39b303fc60c9767ba161b2d2641
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com