Revert "Adding __INTRODUCED_IN to stats_write if contains repeat..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c4883a9e-fb23-4457-92a8-a5eef84088dc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 353432148
Change-Id: I7cc329e3cfdbca65e2a34fcdd2d8e3e83059cf1c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com