Revert "Defining __ANDROID_API_T__ and __INTRODUCED_IN on non-An..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1f55c67c-0e77-4d9e-91f0-127df1266c31). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 353511917
Change-Id: Ic743970940c335a2ff3d2f0149ff90c14497b24b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com