Revert "Adding __INTRODUCED_IN to stats_write if contains repeat..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1f55c67c-0e77-4d9e-91f0-127df1266c31). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 353511917
Change-Id: I215d29c8bb4d1409f4ca1598888b2e670fd9c907
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com