Revert "Add QuirkEvent for PD provisioned"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d5dd2f99-8338-433b-8caa-0a2a68f4e1f6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7dbe1e93ee151e3a878a18172e42b43540f7bd16
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com